+++ src/main/kotlin/kr/hhplus/be/server/presentation/payment/model/PaymentResponse.kt
@@ -1,0 +1,8 @@
+package kr.hhplus.be.server.presentation.payment.model
+
+class PaymentResponse {
+
+    data class Payment(
+        val orderId: Long,
+    )
+}
